% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calcEmStats}
\alias{calcEmStats}
\title{calculate entropy and AUC fo emergence times}
\usage{
calcEmStats(EmTimes, years, EntropyRes)
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

