% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateInflection.R
\name{calcInflectionTimes}
\alias{calcInflectionTimes}
\title{calcInflectionTimes}
\usage{
calcInflectionTimes(tempMat, method, printProgress, ...)
}
\value{
a data.frame
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

