% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergence.R
\name{calculateDomainHorizonTimes}
\alias{calculateDomainHorizonTimes}
\title{calculateDomainHorizonTimes}
\usage{
calculateDomainHorizonTimes(groupName, outFolder, TimeSeriesFolder,
  TempTimeSeries, rangeData, maxT, domain, printProgress, permYearCut, method,
  runLength, prob)
}
\value{
a data.frame
}
\description{
this is a wrapper function for calculating emergence times across the domain for a set of different climate runs and speceis Tmax values
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

