% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{calculateMaxT}
\alias{calculateMaxT}
\title{calculateMaxT}
\usage{
calculateMaxT(groupName, outFolder, TimeSeriesFolder, TempTimeSeries,
  rangeCells, domain, quantWithin = NA, quantAcross = NA, StDevWithin = 3,
  StDevAcross = 3)
}
\value{
a data.frame
}
\description{
calculate the maximum temperature across species geographic ranges
}
\details{
See Examples.
}
\note{
wrapper function for 'extractMaxT' function
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

