% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cm.cols1}
\alias{cm.cols1}
\title{Nice color palettes for ranges}
\usage{
cm.cols1(x, bias = 1)
}
\arguments{
\item{x}{number of colors}

\item{bias}{as in colorRampPalette}
}
\value{
Returns a vector of colors
}
\description{
Nice color palettes for ranges
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
\seealso{
colorRampPalette
}

