% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergence.R
\name{communityEmergenceTimes}
\alias{communityEmergenceTimes}
\title{find emergence times for the species in a cell}
\usage{
communityEmergenceTimes(focmaxT, focSpec, futTs, yearsFut, permYearCut,
  method = "Mora", runLength, prob)
}
\arguments{
\item{focmaxT}{The T max of each species in the cell}

\item{futTs}{The future temperatures in the cell}

\item{yearsFut}{The years for which we have future temperature data}

\item{permYearCut}{The number of years before the end of the projection for defining permanent emergence i.e if the projection runs to 2300 and permYearCut = 10, then all years past 2290 mus be in an emergent state}

\item{method}{Can be Mora or Henson}
}
\value{
a data.frame
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

