% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{CommunityHorizonPlot}
\alias{CommunityHorizonPlot}
\title{Community Horizon Plot}
\usage{
CommunityHorizonPlot(histTs, futTs, yearsHist, yearsFut, focmaxT, focSpec,
  permYearCut, method, runLength, prob, First = FALSE, Perm = FALSE,
  Prob = TRUE, addJitter = TRUE)
}
\arguments{
\item{First}{logical;}

\item{Perm}{logical;}

\item{Prob}{logical;}

\item{addJitter}{logical;}
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

