% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergence.R
\name{domainEmergence}
\alias{domainEmergence}
\title{find emergence times for a set of communities}
\usage{
domainEmergence(domain, rangeData, specT, tempPostMat, printProgress = TRUE,
  permYearCut, method = "Mora", runLength, prob)
}
\arguments{
\item{domain}{Contains the cell ids (WorldID) we want to explore - we could just use a vector these rather than a big dataframe.......}

\item{rangeData}{Dataframe containing species names (Tip_Label) and cell ids (WorldID)}

\item{specT}{Dataframe containing species names (Tip_Label) and max T values}

\item{permYearCut}{The number of years before the end of the projection for defining permanent emergence i.e if the projection runs to 2300 and permYearCut = 10, then all years past 2290 mus be in an emergent state}

\item{futTs}{The future temperatures in the cell}

\item{yearsFut}{The years for which we have future temperature data}

\item{tempFutMat}{Matrix containing future temperature values for each cell in domain}
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

