% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{extractMaxT}
\alias{extractMaxT}
\title{extractMaxT}
\usage{
extractMaxT(ranges, domain, histT, quantWithin, quantAcross, StDevWithin,
  StDevAcross)
}
\arguments{
\item{ranges}{A list with each element containing the cell ids where a species occurs}

\item{domain}{A dataframe containing the cell id's}

\item{histT}{A matrix containing the temperatures for each year (columns) for each cell (row) in 'domain'}

\item{quantWithin}{The quantile to use when calculating the maximum temperature within a cell}

\item{quantAcross}{The quantile to use when calculating the maximum temperature across cells}

\item{StDevWithin}{The number of standard deviations used to exclude extreme temperature values within a cell}

\item{StDevAcross}{The number of standard deviations used to exclude extreme temperature values across the range}
}
\value{
a data.frame
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

