% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateInflection.R
\name{findHensonInflection}
\alias{findHensonInflection}
\title{Find inflections in temperature time series}
\usage{
findHensonInflection(TsData, option = 2, sdDate = 1980, nSD = 2,
  sdThreshold = 1, lowerBound = 1950, upperBound = 2010, verbose = T)
}
\arguments{
\item{TsData}{dataframe?}

\item{option}{1 or 2 for methods to calculate the cumulative gradient}

\item{sdDate}{end data for calculating the SD}

\item{nSD}{nuber of SD that cumulative gradient must excede the mean}

\item{sdThreshold}{distance, in SD, that cumulative gradient must excede the mean}

\item{lowerBound}{year to begin looking for inflection point}

\item{upperBound}{max year to use at inflection point}

\item{verbose}{logical}
}
\value{
a data.frame
}
\description{
Inflection based on change exceding noise
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

