% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateInflection.R
\name{findHensonToE}
\alias{findHensonToE}
\title{findHensonToE}
\usage{
findHensonToE(TsData, option)
}
\value{
a data.frame
}
\description{
function to find the emergence time and also inflection time (i.e. when the temperature time series starts to trend upwards) following the method of Henson (or as close as I can follow it anyway)
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

