% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateInflection.R
\name{findInflectionTimes}
\alias{findInflectionTimes}
\title{findInflectionTimes}
\usage{
findInflectionTimes(domain, inputFolder, TempTimeSeries, method, printProgress)
}
\value{
a data.frame
}
\description{
loads in time series data and produces a matrix of inflection times across the domain
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

