% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergence.R
\name{findMoraToE}
\alias{findMoraToE}
\title{find emergence times for a species}
\usage{
findMoraToE(focmaxT, futTs, yearsFut, runLength, prob)
}
\arguments{
\item{focmaxT}{The T max of each species in the cell}

\item{futTs}{The future temperatures in the cell}

\item{runLength}{The number of consecutive years that temperatures must exceed focmaxT to qualify as emergence}

\item{prob}{The percent of subsequent years that must be emergent to qualify as emergence}
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>,  Cory Merow
}

