% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateInflection.R
\name{findOPTInflection}
\alias{findOPTInflection}
\title{Find inflection time}
\usage{
findOPTInflection(TsData)
}
\value{
a data.frame
}
\description{

}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

