% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mapIt}
\alias{mapIt}
\title{map patterns of emergence}
\usage{
mapIt(gridB, toplot, colors = colorRampPalette(c("steelblue4", "cadetblue2",
  "yellow", "red"))(10), brks2, legendPlot = TRUE)
}
\arguments{
\item{colors}{vector of colors, e.g. from `colorRampPalette`}
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

