% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{NSpeciesEmerging}
\alias{NSpeciesEmerging}
\title{NSpeciesEmerging}
\usage{
NSpeciesEmerging(emergenceTimes, year)
}
\description{
using the species emergence times for each community calculate the number of species emerging by a particular year
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

