% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{splitTimeSeries}
\alias{splitTimeSeries}
\title{splitTimeSeries}
\usage{
splitTimeSeries(InflectionTimes, tempMat)
}
\value{
a data.frame
}
\description{
Takes a matrix with the time series of temperature for each cell and generates two matrices containing the pre and post inflection temperature values
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

