% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summariseHorizonTimes}
\alias{summariseHorizonTimes}
\title{extract emergence metrics for a set of runs}
\usage{
summariseHorizonTimes(groupName, ET, TempTimeSeries, year, yearTr, EntropyRes)
}
\details{
See Examples.
}
\author{
Alex Pigot <alex.pigot1@gmail.com>, Cory Merow
}

